Removing LocalTrackOptions export from types and moved into index.d.ts
